blog/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-Migrate==4.0.7
flask-marshmallow==1.2.1
flask-caching==2.3.0
Flask-Limiter==3.7.0
marshmallow-sqlalchemy==1.1.0
marshmallow==3.21.3
requests==2.32.3
gunicorn==22.0.0
flask-cors==5.0.0
cryptography==44.0.0
click==8.1.8
pyjwt==2.9.0
psycopg2==2.9.10